fix: improper positioning in Safari et al.
In several browsers, getBoundingClientRect returns an object that does
not have x/y properties (see
https://developer.mozilla.org/en-US/docs/Web/API/Element/getBoundingClientRect).

As of 0.1.1, passing the result of getBoundingClientRect to the
constructor would cause a type error.

Since it's very easy to extend the return value to have those
properties, we now do that in our constructor, and update our
constructor's typings to accept ClientRect instead of DOMRect.