Simplify alignment calculations with a scaling function
Just as we define our grid placements on each axis on a [0, 1] domain,
we can do the same for our alignment calculations by mapping our e.g.
[0, width] range to a [0, 1] domain.

This simplifies our enum types as well and smooths the path for defining
any arbitrary alignment, not just the discrete ones we have defined.